Add: module.exports to connect with files
