+++ routes.js
@@ -1,0 +1,43 @@
+const fs = require('fs');
+
+const requestHandler = (req, res) => {
+
+  const url = req.url;
+  const method = req.method;
+
+  if(url === '/') {
+    res.setHeader('Content-Type', 'text/html');
+    res.write('<html>')
+    res.write('<head><title>Enter Message</title></head>')
+    res.write('<body><form action="/message" method="POST"><input type="text" name="message"><button type="submit">Send</button></buttoo></buttin></form></body>')
+    res.write('</html>')
+    return res.end();
+  }
+
+  if(url === '/message' && method === 'POST') {
+    const body = [];
+
+    req.on('data', (chunk) => {
+      body.push(chunk);
+    });
+
+    return req.on('end', () => {
+      const parseBody = Buffer.concat(body).toString();
+      const message = parseBody.split('=')[1];
+      fs.writeFile('message.txt', message, (err) => {
+        res.statusCode = 302;
+        res.setHeader('Location', '/');
+        return res.end();
+      });
+    });
+  }
+
+  res.setHeader('Content-Type', 'text/html');
+  res.write('<html>')
+  res.write('<head><title>My first page</title></head>')
+  res.write('<body><h1>Hello world</h1></body>')
+  res.write('</html>')
+  res.end();
+}
+
+module.exports = requestHandler;
